--- management_api_app/tests/test_api/test_routes/test_ping.py
@@ -1,12 +1,0 @@
-import pytest
-from fastapi import FastAPI
-from httpx import AsyncClient
-from resources import strings
-
-
-pytestmark = pytest.mark.asyncio
-
-
-async def test_ping(app: FastAPI, client: AsyncClient) -> None:
-    response = await client.get(app.url_path_for("ping:get-server-alive"))
-    assert response.json()["message"] == strings.PONG
